+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added [?] array conversion
 * ARRAY Aggregator (http://stackoverflow.com/questions/15887900/group-objects-by-property-in-javascript?rq=1)
 * _ column for whole record
+* SELECT INDEX
+* SELECT RECORDSET
+* SELECT TEXT
 
 ### 0.0.34 (14.12.2014 - 20.12.2014)